requirements.txt
aiohttp==3.8.1
aiosignal==1.2.0
async-timeout==4.0.2
attrs==22.1.0
certifi==2022.6.15
charset-normalizer==2.1.1
discord==1.7.3
discord.py==2.0.0
dnspython==2.2.1
frozenlist==1.3.1
idna==3.3
multidict==6.0.2
pymongo==4.2.0
python-dotenv==0.20.0
requests==2.28.1
urllib3==1.26.11
yarl==1.8.1
